+++ packages/granularity/src/components/GrFormField/index.ts
@@ -1,4 +1,4 @@
 export { default } from './GrFormField.vue'
 export { default as GrFormField } from './GrFormField.vue'
-export { GR_FORM_FIELD_KEY, useGrFormFieldContext, type GrFormFieldContext } from './context'
+export { GR_FORM_FIELD_KEY, type GrFormFieldContext, useGrFormFieldContext } from './context'
 export { grFormFieldConfig } from './config'